task/requirements.txt
psycopg2-binary==2.9.11
gql==3.5.0
aiohttp==3.13.2
python-dotenv==1.0.0
tenacity==8.2.3
pydantic==2.12.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

